fix build #460
ImportError while importing test module /var/lib/jenkins/workspace/e-clms-hrl-vpp_extract_load_stac/tests/test_backend.py.
Hint: make sure your test modules/packages have valid Python names.
Traceback:
/usr/lib64/python3.8/importlib/__init__.py:127: in import_module
    return _bootstrap._gcd_import(name[level:], package, level)
tests/test_backend.py:12: in <module>
    from openeogeotrellis.backend import GpsBatchJobs, GpsProcessing, GeoPySparkBackendImplementation, extract_own_job_info
E   ImportError: cannot import name extract_own_job_info from openeogeotrellis.backend (/var/lib/jenkins/workspace/e-clms-hrl-vpp_extract_load_stac/openeogeotrellis/backend.py)
