+++ tests/test_load_stac.py
@@ -1,0 +1,32 @@
+import datetime as dt
+
+import mock
+import pytest
+from openeo_driver.backend import BatchJobs, BatchJobMetadata
+
+from openeogeotrellis.load_stac import extract_own_job_info
+
+
+@pytest.mark.parametrize("url, user_id, job_info_id",
+                         [
+                             ("https://oeo.net/openeo/1.1/jobs/j-20240201abc123/results", 'alice', 'j-20240201abc123'),
+                             ("https://oeo.net/openeo/1.1/jobs/j-20240201abc123/results", 'bob', None),
+                             ("https://oeo.net/openeo/1.1/jobs/j-20240201abc123/results/N2Q1MjMzODEzNzRiNjJlNmYyYWFkMWYyZjlmYjZlZGRmNjI0ZDM4MmE4ZjcxZGI2Z/095be1c7a37baf63b2044?expires=1707382334", 'alice', None),
+                             ("https://oeo.net/openeo/1.1/jobs/j-20240201abc123/results/N2Q1MjMzODEzNzRiNjJlNmYyYWFkMWYyZjlmYjZlZGRmNjI0ZDM4MmE4ZjcxZGI2Z/095be1c7a37baf63b2044?expires=1707382334", 'bob', None),
+                             ("https://earth-search.aws.element84.com/v1/collections/sentinel-2-l2a", 'alice', None)
+                         ])
+def test_extract_own_job_info(url, user_id, job_info_id):
+    batch_jobs = mock.Mock(spec=BatchJobs)
+
+    def alices_single_job(job_id, user_id):
+        return (BatchJobMetadata(id=job_id, status='finished', created=dt.datetime.utcnow())
+                if job_id == 'j-20240201abc123' and user_id == 'alice' else None)
+
+    batch_jobs.get_job_info.side_effect = alices_single_job
+
+    job_info = extract_own_job_info(url, user_id, batch_jobs=batch_jobs)
+
+    if job_info_id is None:
+        assert job_info is None
+    else:
+        assert job_info.id == job_info_id
